Modyfied:  - fiels  - thread added for encoder  - print out encoder position
