+++ .vscode/settings.json
@@ -1,6 +1,5 @@
 {
     "[python]": {
-        "editor.defaultFormatter": "ms-python.autopep8"
-    },
-    "python.formatting.provider": "none"
+        "editor.defaultFormatter": "ms-python.python"
+    }
 }
